Replace bot.py with simple version for Railway
